feat: extract GameState enum to a separate class
