+++ src/main/java/com/boredxgames/tictactoeclient/domain/model/GameState.java
@@ -1,0 +1,8 @@
+package com.boredxgames.tictactoeclient.domain.model;
+
+public enum GameState {
+    IN_PROGRESS,
+    X_WINS,
+    O_WINS,
+    DRAW
+}
