+++ src/main/java/com/boredxgames/tictactoeclient/domain/services/game/GameService.java
@@ -1,12 +1,13 @@
 package com.boredxgames.tictactoeclient.domain.services.game;
 
+import com.boredxgames.tictactoeclient.domain.model.GameState;
 import com.boredxgames.tictactoeclient.domain.model.Move;
-import com.boredxgames.tictactoeclient.domain.services.GameBoard;
+
 /**
  * @author Tasneem
  */
 public interface GameService {
     void makeMove(Move move, char currentPlayer);
     Move getNextMove(GameBoard board, char currentPlayer);
-    GameBoard.GameState getOutcome(GameBoard board);
+    GameState getOutcome(GameBoard board);
 }
